+++ src/app/api/chat/gemini/route.ts
@@ -1,65 +1,123 @@
-import { type CoreMessage, streamText } from 'ai'
-import { type ChatBody, type Conversation } from '~/types/chat'
+import { type CoreMessage, generateText, streamText } from 'ai'
+import { type Conversation } from '~/types/chat'
 import { createGoogleGenerativeAI } from '@ai-sdk/google'
+import {
+  GeminiModels,
+  type GeminiModel,
+  GeminiModelID,
+} from '~/utils/modelProviders/types/gemini'
+import {
+  type GeminiProvider,
+  ProviderNames,
+} from '~/utils/modelProviders/LLMProvider'
 import { decryptKeyIfNeeded } from '~/utils/crypto'
-import { ProviderNames } from '~/utils/modelProviders/LLMProvider'
-import { GeminiModel, GeminiModels } from '~/utils/modelProviders/types/gemini'
 export const dynamic = 'force-dynamic'
 export const maxDuration = 60
 
 import { NextResponse } from 'next/server'
 
-export async function POST(req: Request) {
+export async function runGeminiChat(
+  conversation: Conversation,
+  geminiProvider: GeminiProvider,
+  stream: boolean,
+) {
+  if (!conversation) {
+    throw new Error('Conversation is missing')
+  }
+
+  if (!geminiProvider.apiKey) {
+    throw new Error('Gemini API key is missing')
+  }
+
   try {
-    const { chatBody }: { chatBody: ChatBody } = await req.json()
-
-    const conversation = chatBody.conversation
-    if (!conversation) {
-      throw new Error('Conversation is missing from the chat body')
-    }
-
-    const apiKey = chatBody.llmProviders?.Gemini?.apiKey
-    if (!apiKey) {
-      throw new Error('Gemini API key is missing')
-    }
-
     const gemini = createGoogleGenerativeAI({
-      apiKey: await decryptKeyIfNeeded(apiKey),
+      apiKey: await decryptKeyIfNeeded(geminiProvider.apiKey),
     })
+
+    if (conversation.messages.length === 0) {
+      throw new Error('Conversation messages array is empty')
+    }
+
+    // Validate model ID
+    if (
+      !Object.values(GeminiModels).some(
+        (model) => model.id === conversation.model.id,
+      )
+    ) {
+      throw new Error(`Invalid Gemini model ID: ${conversation.model.id}`)
+    }
 
     const model = gemini(conversation.model.id)
-    const result = await streamText({
-      model: model as any,
-      messages: convertConversationToVercelAISDKv3(conversation),
-      temperature: conversation.temperature,
-    })
+    console.log('Using Gemini model:', conversation.model.id)
 
-    return result.toTextStreamResponse()
-  } catch (error) {
-    if (
-      error &&
-      typeof error === 'object' &&
-      'data' in error &&
-      error.data &&
-      typeof error.data === 'object' &&
-      'error' in error.data
-    ) {
-      console.error('error.data.error', error.data.error)
-      return new Response(JSON.stringify({ error: error.data.error }), {
-        status: 500,
-        headers: { 'Content-Type': 'application/json' },
-      })
-    } else {
-      return new Response(
-        JSON.stringify({
-          error: 'An error occurred while processing the chat request',
-        }),
-        {
-          status: 500,
-          headers: { 'Content-Type': 'application/json' },
-        },
-      )
+    const messages = convertConversationToVercelAISDKv3(conversation)
+    console.log('Converted messages:', JSON.stringify(messages, null, 2))
+
+    // Check if we're using vision model with image content
+    const hasImageContent = messages.some(
+      (msg) =>
+        typeof msg.content === 'object' &&
+        Array.isArray(msg.content) &&
+        msg.content.some((c) => c.type === 'image'),
+    )
+
+    const commonParams = {
+      model: model as any,
+      messages: messages,
+      temperature: conversation.temperature || 0.7,
+      maxTokens: conversation.model.tokenLimit || 4096,
     }
+    console.log(
+      'Request params:',
+      JSON.stringify(
+        {
+          modelId: conversation.model.id,
+          temperature: commonParams.temperature,
+          maxTokens: commonParams.maxTokens,
+          messageCount: messages.length,
+        },
+        null,
+        2,
+      ),
+    )
+
+    if (stream) {
+      try {
+        const result = await streamText(commonParams)
+        return result.toTextStreamResponse()
+      } catch (error) {
+        console.error('Gemini streaming error:', error)
+        if (
+          error instanceof Error &&
+          error.message.includes('Developer instruction is not enabled')
+        ) {
+          throw new Error(
+            'This Gemini API key does not have access to the requested model. Please verify your API key permissions in the Google AI Studio.',
+          )
+        }
+        throw error
+      }
+    } else {
+      try {
+        const result = await generateText(commonParams)
+        const choices = [{ message: { content: result.text } }]
+        return { choices }
+      } catch (error) {
+        console.error('Gemini generation error:', error)
+        if (
+          error instanceof Error &&
+          error.message.includes('Developer instruction is not enabled')
+        ) {
+          throw new Error(
+            'This Gemini API key does not have access to the requested model. Please verify your API key permissions in the Google AI Studio.',
+          )
+        }
+        throw error
+      }
+    }
+  } catch (error) {
+    console.error('Gemini API error:', error)
+    throw error
   }
 }
 
@@ -104,8 +162,8 @@ function convertConversationToVercelAISDKv3(
   return coreMessages
 }
 
-export async function GET(req: Request) {
-  const apiKey = process.env.ANTHROPIC_API_KEY
+export async function GET() {
+  const apiKey = process.env.GEMINI_API_KEY
 
   if (!apiKey) {
     return NextResponse.json(
